% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseFile.R
\name{chooseFile}
\alias{chooseFile}
\title{Convenient function for choosing CSV data file using interactive dialog box.}
\usage{
chooseFile(...)
}
\arguments{
\item{...}{Any arguments available for read.csv(), read.table(), readxl::read_excel(), foreign::read.spss() functions; file will be automatically chosen once you pick up your file using the file chooser dialog box.}
}
\value{
A character value of the file path will be returned.
}
\description{
You can alternatively use this function for choosing *.csv, *.txt, *.xlsx, *.sav (SPSS) files in an interactive GUI mode A file choose dialog box will be prompted.
}
\details{
chooseFile


See example below.
}
\examples{
data <- chooseFile(header=T)
str(data)
# You can  drop off your file using the file picker dialog box.
}
\author{
JooYoung Seo (jooyoung@psu.edu)
}
\keyword{chooseFile}
